+++ zetaclient/chains/solana/signer/signer.go
@@ -276,11 +276,14 @@ func (signer *Signer) TryProcessOutbound(
 	signer.broadcastOutbound(ctx, outbound, chainID, nonce, logger, zetaRepo)
 }
 
-// signTx creates and signs solana tx containing provided instruction with relayer key.
+// signTx creates and signs a Solana tx containing the provided instruction with the relayer key.
+// If `addressLookupTable` is non-nil and `addrs` is non-empty, the transaction will include an address lookup table.
 func (signer *Signer) signTx(
 	ctx context.Context,
 	inst *sol.GenericInstruction,
 	limit uint64,
+	addressLookupTable *sol.PublicKey,
+	addrs sol.PublicKeySlice,
 ) (*sol.Transaction, error) {
 	// get a recent blockhash
 	recent, err := signer.solanaClient.GetLatestBlockhash(ctx, broadcastOutboundCommitment)
@@ -288,7 +291,7 @@ func (signer *Signer) signTx(
 		return nil, errors.Wrap(err, "getLatestBlockhash error")
 	}
 
-	// if limit is provided, prepend compute unit limit instruction
+	// prepend compute unit limit instruction if needed
 	var instructions []sol.Instruction
 	if limit > 0 {
 		limit = min(limit, SolanaMaxComputeBudget)
@@ -296,17 +299,22 @@ func (signer *Signer) signTx(
 		limitInst := computebudget.NewSetComputeUnitLimitInstruction(uint32(limit)).Build()
 		instructions = append(instructions, limitInst)
 	}
-
 	instructions = append(instructions, inst)
 
-	// create a transaction that wraps the instruction
-	tx, err := sol.NewTransaction(
-		// TODO: outbound now uses 5K lamports as the fixed fee, we could explore priority fee and compute budget
-		// https://github.com/zeta-chain/node/issues/2599
-		instructions,
-		recent.Value.Blockhash,
+	// transaction options
+	opts := []sol.TransactionOption{
 		sol.TransactionPayer(signer.relayerKey.PublicKey()),
-	)
+	}
+
+	// add AddressLookupTable if provided
+	if addressLookupTable != nil && len(addrs) > 0 {
+		opts = append(opts, sol.TransactionAddressTables(map[sol.PublicKey]sol.PublicKeySlice{
+			*addressLookupTable: addrs,
+		}))
+	}
+
+	// create a transaction
+	tx, err := sol.NewTransaction(instructions, recent.Value.Blockhash, opts...)
 	if err != nil {
 		return nil, errors.Wrap(err, "unable to create new tx")
 	}
@@ -380,7 +388,7 @@ func (signer *Signer) broadcastOutbound(
 					break
 				}
 
-				fallbackTx, err := signer.signTx(ctx, fallbackInst, 0)
+				fallbackTx, err := signer.signTx(ctx, fallbackInst, 0, nil, nil)
 				if err != nil {
 					logger.Error().Err(err).Fields(lf).Msg("error signing increment nonce instruction")
 					break
@@ -399,16 +407,18 @@ func (signer *Signer) broadcastOutbound(
 	}
 }
 
-// createOutboundWithFallback is a helper function that creates an outbound with a main and a fallback transaction
-// and signs them with relayer key
+// createOutboundWithFallback creates an outbound with a main transaction and fallback message,
+// signed with the relayer key. If an addressLookupTable is provided, the transaction will include it.
 func (signer *Signer) createOutboundWithFallback(
 	ctx context.Context,
 	mainInst *sol.GenericInstruction,
 	msgIn *contracts.MsgIncrementNonce,
 	computeLimit uint64,
+	addressLookupTable *sol.PublicKey,
+	addrs sol.PublicKeySlice,
 ) (*Outbound, error) {
 	// Create and sign main transaction
-	tx, err := signer.signTx(ctx, mainInst, computeLimit)
+	tx, err := signer.signTx(ctx, mainInst, computeLimit, addressLookupTable, addrs)
 	if err != nil {
 		return nil, errors.Wrap(err, "error signing main instruction")
 	}